+++ src/init.c
@@ -4,13 +4,13 @@
 #include <R_ext/Rdynload.h>
 
 /* .Call calls */
-extern SEXP method_(SEXP, SEXP, SEXP);
+extern SEXP method_(SEXP, SEXP, SEXP, SEXP);
 extern SEXP method_call_(SEXP, SEXP, SEXP);
 extern SEXP S7_class_(SEXP, SEXP);
 extern SEXP S7_object_();
 
 static const R_CallMethodDef CallEntries[] = {
-    {"method_", (DL_FUNC) &method_, 3},
+    {"method_", (DL_FUNC) &method_, 4},
     {"method_call_", (DL_FUNC) &method_call_, 3},
     {"S7_object_", (DL_FUNC) &S7_object_, 0},
     {NULL, NULL, 0}