+++ src/method-dispatch.c
@@ -36,23 +36,54 @@ SEXP method_rec(SEXP table, SEXP signature, R_xlen_t signature_itr) {
   return R_NilValue;
 }
 
-__attribute__ ((noreturn))
-void S7_method_lookup_error(SEXP generic, SEXP signature) {
-  static SEXP S7_method_lookup_error_fun = NULL;
-  SEXP ns = Rf_findVarInFrame(R_NamespaceRegistry, Rf_install("S7"));
+SEXP generic_args(SEXP generic, SEXP envir) {
+  // How many arguments are used for dispatch?
+  SEXP dispatch_args = Rf_getAttrib(generic, Rf_install("dispatch_args"));
+  R_xlen_t n_dispatch = Rf_xlength(dispatch_args);
 
+  // Allocate a list to store the arguments
+  SEXP args = PROTECT(Rf_allocVector(VECSXP, n_dispatch));
+
+  // Find the value of each argument.
+  SEXP formals = FORMALS(generic);
+  for (R_xlen_t i = 0; i < n_dispatch; ++i) {
+    SEXP name = TAG(formals);
+    SEXP arg = Rf_findVar(name, envir);
+
+    if (PRCODE(arg) == R_MissingArg) {
+      SET_VECTOR_ELT(args, i, R_MissingArg);
+    } else {
+      // method_call_() has already done the necessary computation
+      SET_VECTOR_ELT(args, i, Rf_eval(arg, R_EmptyEnv));
+    }
+
+    formals = CDR(formals);
+  }
+  Rf_setAttrib(args, R_NamesSymbol, dispatch_args);
+
+  UNPROTECT(1);
+
+  return args;
+}
+
+__attribute__ ((noreturn))
+void S7_method_lookup_error(SEXP generic, SEXP signature, SEXP envir) {
+  SEXP ns = Rf_findVarInFrame(R_NamespaceRegistry, Rf_install("S7"));
+  static SEXP S7_method_lookup_error_fun = NULL;
   if (S7_method_lookup_error_fun == NULL) {
     S7_method_lookup_error_fun = Rf_findVarInFrame(ns, Rf_install("method_lookup_error"));
   }
+
   SEXP name = Rf_getAttrib(generic, R_NameSymbol);
-  SEXP args = Rf_getAttrib(generic, Rf_install("dispatch_args"));
+  SEXP args = generic_args(generic, envir);
+
   SEXP S7_method_lookup_error_call = PROTECT(Rf_lang4(S7_method_lookup_error_fun, name, args, signature));
   Rf_eval(S7_method_lookup_error_call, ns);
 
   while(1);
 }
 
-SEXP method_(SEXP generic, SEXP signature, SEXP error_) {
+SEXP method_(SEXP generic, SEXP signature, SEXP envir, SEXP error_) {
   if (!Rf_inherits(generic, "S7_generic")) {
     return R_NilValue;
   }
@@ -66,7 +97,7 @@ SEXP method_(SEXP generic, SEXP signature, SEXP error_) {
 
   int error = Rf_asInteger(error_);
   if (error && m == R_NilValue) {
-    S7_method_lookup_error(generic, signature);
+    S7_method_lookup_error(generic, signature, envir);
   }
 
   return m;
@@ -160,7 +191,7 @@ SEXP method_call_(SEXP call, SEXP generic, SEXP envir) {
   }
 
   // Now that we have all the classes, we can look up what method to call
-  SEXP m = method_(generic, dispatch_classes, Rf_ScalarLogical(1));
+  SEXP m = method_(generic, dispatch_classes, envir, Rf_ScalarLogical(1));
   SETCAR(mcall, m);
 
   UNPROTECT(n_protect);