+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## March 2023
 
+* Improve method dispatch failure message (#231).
+
 * Can use `|` to create unions from S7 classes (#224).
 
 * Can no longer subclass an environment via `class_environment` because we